refactor(menu)!: extract drawing logic to MenuRenderer
